Moved loginmiddleware into middleware folder
