+++ src/middleware/loginmiddleware.js
@@ -1,6 +1,6 @@
-import {SocketConfig, logger, ErrorMessage} from './conf/config.js'
-import {DataBaseManager} from "./database/DatabaseManager.js";
-import {dbManager} from "./database/DatabaseManager.js";
+import {SocketConfig, logger, ErrorMessage} from '../conf/config.js'
+import {DataBaseManager} from "../database/DatabaseManager.js";
+import {dbManager} from "../database/DatabaseManager.js";
 
 /**
  * Class implementing a Middleware function with a fixed password
